:truck: Move view files
Move to characters, episodes and favorites folders
