+++ pages/_app.tsx
@@ -5,7 +5,7 @@ import { ThemeProvider } from "@mui/material/styles";
 import { theme } from "../src/themes/mainThemeMUI";
 import { ApolloProvider } from "@apollo/client";
 import client from "../apollo-client";
-import Layout from "../src/components/layout";
+import Layout from "../src/components/layouts/mainLayout";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
